src/requirements.txt
aiocron==1.8
aiogram==3.17.0
aiohttp==3.11.0
asyncio==3.4.3
ciso8601==2.3.1
orjson==3.9.10
python-dotenv==1.0.0
pillow==11.1.0
